// src/types/api.ts
export interface ApiReturn<T> {
  data: T;
  message: string;
  success: number;
}

export interface ApiError {
  code?: number;
  message: string;
  status: number;
}

export type UninterceptedApiError = {
  message: string | Record<string, string[]>;
};

type PaginateData<Data> = {
  data_per_page: Data;
  meta: {
    page: number;
    max_page: number;
  };
};

export interface PaginatedApiResponse<DataType> {
  code: number;
  success: string;
  data: PaginateData<DataType>;
}

export interface PaginatedApiReturn<DataType> {
  code: number;
  success: string;
  data: DataType;
  metadata: {
    total: number;
    page: number;
    limit: number;
    page_count: number;
  };
}